sql: refactor unimplemented statements into one file pkg/sql/unimplemented.go
Moved files with unimplemented stubs into just 1 file unimplemented.go.
These unimplemented stubs work in DSC only and we will not implement
them on the legacy schema changer.

Fixes: #137615
Release note: none
