+++ pkg/sql/BUILD.bazel
@@ -13,7 +13,6 @@ go_library(
         "alter_function.go",
         "alter_index.go",
         "alter_index_visible.go",
-        "alter_policy.go",
         "alter_primary_key.go",
         "alter_role.go",
         "alter_schema.go",
@@ -41,7 +40,6 @@ go_library(
         "comment_on_index.go",
         "comment_on_schema.go",
         "comment_on_table.go",
-        "comment_on_type.go",
         "compact_sql_stats.go",
         "completions.go",
         "conn_executor.go",
@@ -64,7 +62,6 @@ go_library(
         "create_external_connection.go",
         "create_function.go",
         "create_index.go",
-        "create_policy.go",
         "create_role.go",
         "create_schema.go",
         "create_sequence.go",
@@ -102,14 +99,11 @@ go_library(
         "drop_external_connection.go",
         "drop_function.go",
         "drop_index.go",
-        "drop_owned_by.go",
-        "drop_policy.go",
         "drop_role.go",
         "drop_schema.go",
         "drop_sequence.go",
         "drop_table.go",
         "drop_tenant.go",
-        "drop_trigger.go",
         "drop_type.go",
         "drop_view.go",
         "error_hints.go",
@@ -279,6 +273,7 @@ go_library(
         "txn_state.go",
         "type_change.go",
         "unary.go",
+        "unimplemented.go",
         "union.go",
         "unlisten.go",
         "unsplit.go",
