--- pkg/sql/drop_owned_by.go
@@ -1,27 +1,0 @@
-// Copyright 2020 The Cockroach Authors.
-//
-// Use of this software is governed by the CockroachDB Software License
-// included in the /LICENSE file.
-
-package sql
-
-import (
-	"context"
-
-	"github.com/cockroachdb/cockroach/pkg/server/telemetry"
-	"github.com/cockroachdb/cockroach/pkg/sql/sqltelemetry"
-	"github.com/cockroachdb/cockroach/pkg/util/errorutil/unimplemented"
-)
-
-func (p *planner) DropOwnedBy(ctx context.Context) (planNode, error) {
-	if err := checkSchemaChangeEnabled(
-		ctx,
-		p.ExecCfg(),
-		"DROP OWNED BY",
-	); err != nil {
-		return nil, err
-	}
-	telemetry.Inc(sqltelemetry.CreateDropOwnedByCounter())
-	// TODO(angelaw): Implementation.
-	return nil, unimplemented.NewWithIssue(55381, "drop owned by is not yet implemented")
-}
